+++ app/src/main/java/com/jeromedusanter/aircalltest/ui/main/features/repogithub/list/filter/RepoGithubFilterUiModel.kt
@@ -4,6 +4,6 @@ import com.jeromedusanter.aircalltest.ui.base.IUiModel
 
 data class RepoGithubFilterUiModel(
     val sort: RepoGithubSortUiModel,
-    val perPage: Long,
+    val perPage: Int,
     val query: String
 ) : IUiModel
